pyproject.toml
[tool.poetry]
name = "auction_demo"
version = "0.1.0"
description = ""
authors = ["Ruslan Mamleev <mamleev.rus@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.1.1"
djangorestframework = "3.11.1"
black = "19.10b0"
flake8 = "^3.8.3"
celery = "^5.0.0"
django-3-jet = "^1.0.8"
psycopg2-binary = "^2.8.6"
drf-yasg = "^1.17.1"
PyJWT = "^1.7.1"
redis = "^3.5.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
